mysql_pandas/requirements.txt
click==7.1.2
mysql-connector-python==8.0.23
numpy==1.20.1
pandas==1.2.3
pip-tools==5.5.0
pkg-resources==0.0.0
protobuf==3.15.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
